changing to using checkmark
